backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2
pandas
numpy
scikit-learn
torch
transformers
tensorflow
xgboost
prophet
mlxtend
requests
python-dotenv
